+++ tests/phpunit/tests/parentless-categories.php
@@ -76,6 +76,12 @@ class Parentless_Categories_Test extends WP_UnitTestCase {
 	//
 
 
+	public function test_hooks_plugins_loaded() {
+		$this->assertEquals( 10, has_action( 'c2c_parentless_categories', 'c2c_parentless_categories' ) );
+		$this->assertEquals( 10, has_action( 'c2c_get_parentless_categories_list', 'c2c_get_parentless_categories_list' ) );
+		$this->assertEquals( 10, has_action( 'c2c_get_parentless_categories', 'c2c_get_parentless_categories' ) );
+	}
+
 	/* c2c_get_parentless_categories() */
 	public function test_post_with_all_categories_in_branch_assigned_for_c2c_get_parentless_categories() {
 		$post_id = $this->factory->post->create();