+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * New: Add `composer.json` for PHPUnit Polyfill dependency
     * Hardening: Prevent direct web access to `bootstrap.php`
     * Change: In bootstrap, store path to plugin directory in a constant
+    * New: Add tests for function invocation filters
 
 ## 2.2.1 _(2023-05-20)_
 * Change: Note compatibility through WP 6.3+